requirements.txt
asgiref==3.5.2
astroid==2.11.7
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dill==0.3.5.1
Django==4.0.6
django-phonenumber-field==6.3.0
djangorestframework==3.13.1
Faker==13.15.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.23.1
pandas==1.4.3
pathspec==0.9.0
phonenumbers==8.12.51
platformdirs==2.5.2
psycopg2==2.9.3
pylint==2.14.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.1
trycourier==4.2.0
typing_extensions==4.3.0
urllib3==1.26.10
wrapt==1.14.1
